requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
randomname = "^0.1.3"
requests = "^2.25.1"
numpy = "^1.19.5"
Pillow = "^8.3.2"
skl2onnx = "^1.9.3"
onnxruntime = "1.9.0"
tabulate = "0.8.9"
torch = "^1.9.0"
torchvision = "0.10.0"
python-json-logger = "2.0.2"
